+++ traefik.tf
@@ -11,6 +11,8 @@ module "generic" {
     {
       deployment           = var.deployment
       ingress_routes_tcp   = var.ingress_routes_tcp
+      logs                 = var.logs
+      metrics              = var.metrics
       ports                = var.ports
       name                 = var.helm_release.name
       namespace            = var.namespace.name