+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- markdownlint-configure-file { "MD024": { "allow_different_nesting": true } } -->
 
+## [0.13.0] 2023-09-14
+
+### Added
+
+- Input variable `metrics` to configure metrics
+- Input variable `logs` to configure logs
+
+### Changed
+
+- Bump chart version from 17.0.5 to 20.8.0
+
 ## [0.12.0] 2023-08-02
 
 ### Added
@@ -186,6 +197,7 @@ attribute `from_non_www_to_www` and `from_www_to_non_www`
 
 - Terraform module creation
 
+[0.13.0]: https://usine.solution-libre.fr/french-high-availability-multi-cloud-hosting/terraform-modules/traefik/-/compare/v0.12.0...v0.13.0
 [0.12.0]: https://usine.solution-libre.fr/french-high-availability-multi-cloud-hosting/terraform-modules/traefik/-/compare/v0.11.0...v0.12.0
 [0.11.0]: https://usine.solution-libre.fr/french-high-availability-multi-cloud-hosting/terraform-modules/traefik/-/compare/v0.10.1...v0.11.0
 [0.10.1]: https://usine.solution-libre.fr/french-high-availability-multi-cloud-hosting/terraform-modules/traefik/-/compare/v0.10.0...v0.10.1